requirements.txt
beautifulsoup4==4.12.3
flask==3.0.3
openai==1.61.1
python-dotenv==1.0.1
simplegmail==4.1.1
